05_query_analysis/requirements.txt
# langchain==0.1.19
# langchain-chroma==0.1.0
# langchain-community==0.0.38
# langchain-openai==0.1.6

langchain==0.1.19
langchain-chroma==0.1.0
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.15
youtube-transcript-api==0.6.2
pytube==15.0.0
python-dotenv==1.0.1
asyncio==3.4.3